ApiWrapper electron refactor
initial refactor to change Api/db wrapper depending on env, whether in browser or electron - not finished, tested that it can detect electron env even in development served from server
